feat: move function to employee.py
