+++ time_capture/scripts/employee.py
@@ -1,0 +1,18 @@
+import frappe
+from frappe import _
+from frappe.utils import getdate
+from datetime import date
+
+def create_leave_policy_assignment(doc, method=None):
+	
+	year = getdate(doc.creation).year
+	leave_allocation = frappe.new_doc("Leave Policy Assignment")
+	leave_allocation.employee = doc.employee
+	leave_allocation.leave_policy = doc.leave_policy
+	leave_allocation.effective_from = date(year, 1, 1)
+	leave_allocation.effective_to   = date(year, 12, 31)
+	leave_allocation.insert()
+	leave_allocation.submit()
+
+def after_save(doc, method=None):
+	create_leave_policy_assignment(doc, method)
